requirements.txt
# ============================================================
# THƯ VIỆN CẦN THIẾT CHO DỰ ÁN
# Hệ thống phân tích và dự đoán giá cổ phiếu
# ============================================================

# ----------------------------------------------------------
# Web Framework
# ----------------------------------------------------------
streamlit>=1.28.0

# ----------------------------------------------------------
# Xử lý dữ liệu
# ----------------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0

# ----------------------------------------------------------
# Biểu đồ và trực quan hóa
# ----------------------------------------------------------
plotly>=5.18.0

# ----------------------------------------------------------
# Cơ sở dữ liệu
# ----------------------------------------------------------
psycopg2-binary>=2.9.9

# ----------------------------------------------------------
# Machine Learning
# ----------------------------------------------------------
scikit-learn>=1.3.0

# ----------------------------------------------------------
# Dự đoán chuỗi thời gian
# ----------------------------------------------------------
statsmodels>=0.14.0
pmdarima>=2.0.3

# ----------------------------------------------------------
# Các thư viện hỗ trợ khác (tùy chọn)
# ----------------------------------------------------------
# tensorflow>=2.13.0      # Cho LSTM (nếu train model)
# keras>=2.13.0           # Cho LSTM (nếu train model)
# prophet>=1.1.4          # Cho Facebook Prophet (nếu cần)

